query builder generation only from connection and not session fix SchemaBuilder typo fix insertUsing parameters types
