+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "ludb",
-    "version": "0.0.8",
+    "version": "0.0.9",
     "description": "Nodejs Query Builder",
     "author": "Claudio Pennati <claudio.pennati@gmail.com>",
     "license": "MIT",
